upcontacts_api/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
psycopg[binary]>=3.1.16
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic[email]>=2.5.3
python-dotenv>=1.0.0
redis>=5.0.1
slowapi>=0.1.9
cloudinary>=1.38.0
fastapi-mail>=1.4.1
alembic>=1.13.1
pillow>=10.2.0
httpx>=0.26.0
aiofiles>=23.2.1
pydantic-settings>=2.1.0